# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VidVault AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos/{videoId}/ai/generate:
    post:
      operationId: generateAiContent
      tags: [ai]
      summary: Generate AI content for a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAiContentRequest"
      responses:
        "200":
          description: Generated AI content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiOutput"
  /videos/{videoId}/ai/outputs:
    get:
      operationId: listAiOutputs
      tags: [ai]
      summary: List AI outputs for a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of AI outputs
          content:
            application/json:
              schema:
                type: object
                properties:
                  outputs:
                    type: array
                    items:
                      $ref: "#/components/schemas/AiOutput"
                required:
                  - outputs
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI about a video or general queries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
...
components:
  schemas:
...
    AiOutputType:
      type: string
      enum:
        - summary
        - notes
        - ppt_outline
        - mcq
        - flashcards
        - blog_article
        - key_insights
    AiOutput:
      type: object
      properties:
        id:
          type: string
        videoId:
          type: string
        type:
          $ref: "#/components/schemas/AiOutputType"
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - videoId
        - type
        - content
        - createdAt
    GenerateAiContentRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/AiOutputType"
      required:
        - type
    AiChatRequest:
      type: object
      properties:
        message:
          type: string
        videoId:
          type: string
          nullable: true
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...