# lib/api-zod/src/generated/types/aiChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiChatRequestHistoryItem } from "./aiChatRequestHistoryItem";

export interface AiChatRequest {
  message: string;
  videoId?: string | null;
  history?: AiChatRequestHistoryItem[];
}