# lib/api-zod/src/generated/types/aiOutputType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */

export type AiOutputType = (typeof AiOutputType)[keyof typeof AiOutputType];

export const AiOutputType = {
  summary: "summary",
  notes: "notes",
  ppt_outline: "ppt_outline",
  mcq: "mcq",
  flashcards: "flashcards",
  blog_article: "blog_article",
  key_insights: "key_insights",
} as const;