# lib/api-zod/src/generated/types/generateAiContentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiOutputType } from "./aiOutputType";

export interface GenerateAiContentRequest {
  type: AiOutputType;
}